+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.10] - 2025-06-24
+
+### Added
+
+- Added `randomBytes` property to `Cipher` interface and `AesCipher` class
+  - The property is public and readonly, providing access to the random bytes generation function
+  - Updated README to document the new interface property
+
 ## [0.1.9] - 2025-06-09
 
 ### Added